Refactor Resumen Ejecutivo, export & UI
Introduce several improvements across the executive-summary pipeline and UI: cache and inject FechaDe selector into ResumenEjecutivoCalculator to avoid repeated date extraction and pass it through helper methods; change AgenciasDesaparecidas to AgenciasDesaparecidasResumenDto (totals + Top30) and update callers (SlidesPresentacionCalculator, Word exporter, UI) to use totals and top list; harden Word exporter (HTML-encode text, add tree/frequency tables, limit depth, formatting helpers); make GraficaTendencia support money vs p/km JS formatters; improve ArbolJerarquia sorting (multi-column keys, visual active class) and ConsultaViajes filtering to avoid unnecessary ToList allocations; change MetricasMes.Viajes from int to decimal and stop rounding until display. These changes improve performance, correctness, and safety of exports and UI behaviors.

+++ Testing/Features/PivoteViajesModels.cs
@@ -2,7 +2,7 @@
 
 public sealed record MesColumna(string Clave, string Etiqueta);
 
-public sealed record MetricasMes(int Viajes, decimal Km, decimal Peaje, decimal Venta)
+public sealed record MetricasMes(decimal Viajes, decimal Km, decimal Peaje, decimal Venta)
 {
     public decimal KmPorViaje => Viajes > 0 ? Km / Viajes : 0;
     public decimal VentaPorKm => Km > 0 ? Venta / Km : 0;
